feat(1729); Adjustments after code review
